+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Changes
 * [UI/Developer]: Updated to jquery v3.4.0 to fix security issue.
 * [Database]: Fixed issue where FastQC description was being stored with an invalidly formatted version in the database.
 * [UI]: Fixed bug causing issues with saving Line List templates.
+* [UI]: Fixed bug when selecting all samples on the project samples page would not add them to cart.
 
 0.22.0 to 19.01
 ----------------